Separate role-menu Discord transport, validation, and presentation
